renamed docstring in keys.py
