--- Module/dates.py
@@ -1,4 +1,0 @@
-from datetime import date
-
-today = date.today()
-current_year = today.year
