+++ run.py
@@ -1,8 +1,9 @@
 from gspread.exceptions import SpreadsheetNotFound
 
-from testing.delete_spreadsheet import delete_spreadsheet
 
 from Module.template import template
+
+from Module.tests import delete_spreadsheet
 
 
 def main():
@@ -13,7 +14,7 @@ def main():
 
     # Deleted spreadsheet
     delete_spreadsheet(
-        file_id="1tvAopLt5L_IJmsvviDA4SXHa_n5jL8pVNb43H2tLHgI",
+        file_id="1ZK1dHDCXaloQzAnrY6ZlCskypVKRCqUTt48sFXnI21g",
     )
 
     # check to see if the spreadsheet with current_year is created
